+++ server/tests/test_explicit_query.py
@@ -1,0 +1,4 @@
+from test_helper import *
+
+if __name__ == "__main__":
+    test_on_text("Hey Convoscope, do cars generally have 4 wheels?")
