+++ server/tests/test_on_lex.py
@@ -3,6 +3,7 @@ import glob
 import webvtt
 from datetime import datetime
 import random
+from test_helper import *
 
 def load_lex_transcripts(random_n = None, transcript_folder = "./lex_whisper_transcripts", chunk_time_seconds=120):
     podcast_transcript_chunks = dict() # dict of lists
@@ -29,9 +30,22 @@ def load_lex_transcripts(random_n = None, transcript_folder = "./lex_whisper_tra
                 podcast_transcript_chunks[convo_name][-1] += caption.text
     return podcast_transcript_chunks
     
+def test_using_lex_transcript():
+    print("=== Begin Transcript Uploading ===")
+    lex = test_on_lex.load_lex_transcripts(random_n=1)
+    for transcript in lex.keys():
+        for chunk in lex[transcript]:
+            # First, send as intermediate 
+            test_on_text(chunk)
+        print("=== Transcript Uploading Complete ===")
+
+    ui_poll_loop()
+
 if __name__ == "__main__":
-    print(load_lex_transcripts(random_n=3))
-    d = load_lex_transcripts(random_n=3)
-    print(d.keys())
-    for l in d.keys():
-        print(d) 
+    # print(load_lex_transcripts(random_n=3))
+    # d = load_lex_transcripts(random_n=3)
+    # print(d.keys())
+    # for l in d.keys():
+    #    print(d) 
+
+    test_using_lex_transcript()
